Fixed build errors and warnings, found and marked run-time error with project and rename...

+++ Project2-I/Project2-I/interpreter.cpp
@@ -75,7 +75,9 @@ void interpreter::answerQuery(predicate &queryToAnswer)
 	tempRelation.project();
 	//using project function, reduce columns in indexes specified
 	//rename columns to list of project variables
+
 	tempRelation.rename();
+	//PROBLEM HERE ^^^
 
 	//we now have the final relation result for this query. We need to do an insert into the unordered map with the key as the whole query string and the relational result as the mapped value. Since running the same query twice should produce the same results, it doesn't matter if an insert is ignored if run again.
 	relationResults.insert(pair<string, relation>(queryMapKeyGenerator(queryToAnswer), tempRelation));
@@ -204,7 +206,7 @@ void interpreter::printTuples(predicate &QueryToFetchResults)
 
 void interpreter::printTuple(Tuple tupleToPrint, scheme schemeToPrint)
 {
-	for (int i = 0; i < schemeToPrint.size(); i++)
+	for (size_t i = 0; i < schemeToPrint.size(); i++)
 	{
 		queryResults << "  " << schemeToPrint[i] << "='" << tupleToPrint[i] << "'";
